{{ cookiecutter.project_slug }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_name_dashed }}"
version = "0.1.0"
description = ""
authors = ["{{ cookiecutter.author_name }} <{{ cookiecutter.author_email }}>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-allauth = "^0.42.0"
django-anymail = {extras = ["mailgun"], version = "^8.0"}
django-storages = "^1.10.1"
django-environ = "^0.4.5"
django-on-heroku = "^1.0.1"
celery = "4.4.7"
django-celery-results = "^1.2.1"
django-celery-beat = "^2.0.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
